we used Henry's code to make changes and tested... the horn patterns

+++ src/main/java/frc/robot/subsystems/SubsystemCannons.java
@@ -79,7 +79,7 @@ public class SubsystemCannons extends SubsystemBase {
         }
     }
 
-    public void fireHorn(int pulseTime)
+    public void fireHorn(double pulseTime)
     {
         solenoidHorn.setPulseDuration(pulseTime);
         solenoidHorn.startPulse();
